Development (#2)
* Added device monitor functionality and permissions

Added the ability to register a device state monitor that alerts
you when the device is connected, disconnected, and has runtime
permissions granted.

The Sensors interface now throws an exception to the caller
if the device is not physically connected.

Also updated Kotlin dependencies in the readme.md.

* Changing sensor initialization timing

Changed the initialization function to check the return value of the
set operation and the write. Eliminated the flatMapConcat. Also added
a hack to delay between sensors.

Also implemented disconnect logic to join the USB reader coroutine and
use synchronization blocks to prevent problems.

Also:
  Added more debugging.
  Made the argument to checkPermission be nullable.
  Use consistent USB definitions from the M400C code for better readability.

* Remapping device axes

The device axes match reference phones in the documented Android
orientation for
 o  TYPE_ACCELEROMETER
 o  TYPE_GYROSCOPE
 o  TYPE_MAGNETIC_FIELD

AD-1644

* Updated sensor conversions to Android standard

Now reads the HID configuration and HID report at power-on to
dump them to the debug window. Hard-coded into the conversion
routines the values reported by an M400C HID report.

Implemented quaternion rotation to adjust the axes to match
standard Android definition.

Reads back the setup from the device to be sure we started the
sensor. Fails if the data comes back wrong.

AD-1603

* Only initialize selected sensors

Users may only be interested in a subset of sensors, so reduce
USB traffic and power consumption by only initializing the
ones that are requested.

Also fixed pending intent for Android 12 by specifying mutable.

* Fixed units on magnetic flux to be micro teslas

* Wait for sensors to send data upon initialization

Since it is possible some sensors do not start, we now wait for
each to send data before considering it started. If the requested
sensor fails to send data we generate an error and stop listening.

AD-1589

* Bumped revision to 0.0.2 and updated readme file

Preparing for merge from development to master and an updated
version being made public.

* Sensors class code cleanup.

* Wrap some initialization setup in a Debug check.
Rename initSensor to setSensorState since it both stops and starts
a sensor.

* Minor refactor of quaternion rotations

AD-1666

* Updated HID from 458 to 509

Bumped revision to 0.0.3

Co-authored-by: John Westendarp