Added connect retry and extra permission checks for sensors
When the device is connected while running, our activity goes
through stop/destroy/create cycle. Then it misses any broadcast.
If we immediately check, we don't find it because it is not
finished enumerating. We solve this by starting a timer when we
first register and if we miss the broadcast we find it after
it finishes enumerating.

Added automatic permission query when we ask for sensors
interface, too.

Added more debugging and created a consistent debug tag.